First Basic PyTorch Model, Linear Regression
Using L1Loss (Mean Squared Error) as a loss function to train a basic PyTorch model to fit a linear equation. This data was generated by myself, and it was a very simple purpose, with the main goal of this program being to hammer in fundamentals of training loops, and steps to test a model. there are also steps on saving models to a new directory, or to the current directory, and loading saved models.